+++ packages/factory/build.config.ts
@@ -1,5 +1,0 @@
-import { defineConfig } from '@enochfe/cli'
-
-export default defineConfig({
-  entry: 'src/index.ts'
-})