+++ packages/factory/src/unplugin/constants.ts
@@ -1,4 +1,0 @@
-export const DEFINE_FACTORY = 'defineFactory'
-
-export const REGEX_SETUP_SFC = /\.setup\.[cm]?[jt]sx?$/
-export const REGEX_VUE_SFC = /\.vue$/