+++ packages/uncode/src/config.ts
@@ -1,8 +1,0 @@
-export interface FactoryConfig {
-  el: string
-  [index: string]: any
-}
-
-export const definefactory = (config: FactoryConfig) => {
-  return config
-}