+++ packages/uncode/src/hooks/use-factory.ts
@@ -1,6 +1,0 @@
-export const useFactory = () => {
-  const render: any[] = []
-  return {
-    render
-  }
-}