+++ play/src/main.ts
@@ -1,28 +1,34 @@
 import { createApp } from 'vue'
-import { createRouter, createWebHashHistory } from 'vue-router'
-import { setupLayouts } from 'virtual:generated-layouts'
-import routes from '~pages'
-import 'uno.css'
+import { createRouter, createWebHashHistory } from 'vue-router/auto'
 
-import Factory from '@enochfe/factory'
+import { ElNotification } from 'element-plus'
+
+import { createFactory } from '@enochfe/factory'
+
+import '@unocss/reset/tailwind.css'
+import 'uno.css'
 
 import App from './app.vue'
 
-export const router = createRouter({
-  routes: setupLayouts(routes),
-  history: createWebHashHistory()
+import 'element-plus/theme-chalk/src/message.scss'
+import 'element-plus/theme-chalk/src/loading.scss'
+import 'element-plus/theme-chalk/src/notification.scss'
+import 'element-plus/theme-chalk/src/message-box.scss'
+
+export const router = createRouter({ history: createWebHashHistory() })
+
+const factory = createFactory({
+  ajax: {
+    async interceptor(data) {
+      if (data.errors && data.errors.length) {
+        if (!data.errors[0].shouldNotNotification) {
+          ElNotification({ title: '请求失败', message: data.errors[0].message })
+        }
+      }
+
+      return Promise.resolve(data)
+    }
+  }
 })
 
-createApp(App)
-  .use(router)
-  .use(
-    Factory({
-      paging: {
-        itemCount: 0,
-        pageCount: 0,
-        pageIndex: 1,
-        pageSize: 20
-      }
-    })
-  )
-  .mount('#app')
+createApp(App).use(router).use(factory).mount('#app')